Impelement fi_recv() variants
-Modify fi_recv() to use generic _gnix_recv() function.
-Implement fi_recvv(), fi_recvmsg().
-Add tests for each new interface to test/rdm_sr.c.

Fixes ofi-cray/libfabric-cray#244.
