+++ prov/gni/include/gnix.h
@@ -116,7 +116,8 @@ extern "C" {
 
 #define GNIX_SUPPRESS_COMPLETION	(1ULL << 60)	/* TX only flag */
 #define GNIX_RMA_RDMA			(1ULL << 61)	/* RMA only flag */
-#define GNIX_MSG_RENDEVOUS		(1ULL << 61)	/* MSG only flag */
+#define GNIX_MSG_RENDEZVOUS		(1ULL << 61)	/* MSG only flag */
+#define GNIX_MSG_TAGGED			(1ULL << 62)	/* MSG only flag */
 
 /*
  * Cray gni provider supported flags for fi_getinfo argument for now, needs
